Upgraded from C to C++ versions of standard functions

+++ aspects/fluid/conductor/test/UtGunnsFluidCondensingHx.cpp
@@ -311,14 +311,14 @@ void UtGunnsFluidCondensingHx::testHtc()
     tArticle->mHtcCoeff0   = 100.0;
     tArticle->mHtcCoeff1   = 1.0E10;
     tArticle->mHtcLimit    = 2000.0;
-    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * powf(mdot, 20.0))) * 0.5 / tHxNumSegments;
+    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * std::pow(mdot, 20.0))) * 0.5 / tHxNumSegments;
     tArticle->computeHeatTransferCoefficient();
     tArticle->degradeHeatTransferCoefficient();
     CPPUNIT_ASSERT_DOUBLES_EQUAL(expectedHtc,   tArticle->mSegmentHtc, DBL_EPSILON);
 
     mdot = 10.0;
     tArticle->mFlowRate = mdot;
-    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * powf(mdot, 20.0))) * 0.5 / tHxNumSegments;
+    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * std::pow(mdot, 20.0))) * 0.5 / tHxNumSegments;
     tArticle->computeHeatTransferCoefficient();
     tArticle->degradeHeatTransferCoefficient();
     CPPUNIT_ASSERT_DOUBLES_EQUAL(expectedHtc,   tArticle->mSegmentHtc, DBL_EPSILON);
@@ -326,14 +326,14 @@ void UtGunnsFluidCondensingHx::testHtc()
     mdot = FLT_EPSILON;
     tArticle->mFlowRate    = mdot;
     tArticle->mHtcExponent = 0.05;
-    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * powf(mdot, 20.0))) * 0.5 / tHxNumSegments;
+    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * std::pow(mdot, 20.0))) * 0.5 / tHxNumSegments;
     tArticle->computeHeatTransferCoefficient();
     tArticle->degradeHeatTransferCoefficient();
     CPPUNIT_ASSERT_DOUBLES_EQUAL(expectedHtc,   tArticle->mSegmentHtc, DBL_EPSILON);
 
     mdot = 10.0;
     tArticle->mFlowRate = mdot;
-    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * powf(mdot, 20.0))) * 0.5 / tHxNumSegments;
+    expectedHtc = std::min(2000.0, (100.0 + 1.0E10 * std::pow(mdot, 20.0))) * 0.5 / tHxNumSegments;
     tArticle->computeHeatTransferCoefficient();
     tArticle->degradeHeatTransferCoefficient();
     CPPUNIT_ASSERT_DOUBLES_EQUAL(expectedHtc,   tArticle->mSegmentHtc, DBL_EPSILON);
